==> notebooks/05_catogory_identifying.ipynb <==
# Import libraries
import pandas as pd

df_b = pd.read_csv('../datasets/business4.csv')

df_b.info()

df_b[['business_id', 'categories']].head()

# -----------------------------
# 1. Define Category Mapping
# -----------------------------
# Each key represents a consolidated category label (feature column name),
# and the value is a list of raw Yelp category strings associated with that group.
# This mapping will later be used to assign each business to one or more standardized categories.

category_map = {
    "c_nightlife": [
        "Bars", "Lounges", "Pubs", "Sports Bars", "Dive Bars", "Cocktail Bars",
        "Whiskey Bars", "Wine Bars", "Beer Bar", "Champagne Bars", "Hookah Bars",
        "Cigar Bars", "Tiki Bars", "Speakeasies", "Nightlife"
    ],
    "c_cafes": [
        "Cafes", "Coffee & Tea", "Coffee Roasteries", "Themed Cafes",
        "Internet Cafes", "Parent Cafes", "Hong Kong Style Cafe", "Bubble Tea"
    ],
    "c_brunch": [
        "Breakfast & Brunch", "Pancakes", "Waffles", "Bagels", "Cafeteria", "Diners"
    ],
    "c_sandwich": [
        "Sandwiches", "Delis", "Delicatessen", "Cheesesteaks", "Hot Dogs", "Wraps", "Pretzels"
    ],
    "c_american_trad": [
        "American (Traditional)", "Soul Food", "Southern", "Comfort Food",
        "Barbeque", "BBQ", "Steakhouses"
    ],
    "c_american_new": [
        "American (New)", "Gastropubs", "Modern European", "New American"
    ],
    "c_italian": [
        "Italian", "Pasta Shops", "Tuscan", "Sicilian", "Roman", "Pizza"
    ],
    "c_burgers": [
        "Burgers", "Fast Food (Burger Chains)", "Drive-Thru Bars"
    ],
    "c_mexican": [
        "Mexican", "Tex-Mex", "Tacos", "New Mexican Cuisine", "Latin American"
    ],
    "c_seafood": [
        "Seafood", "Fish & Chips", "Seafood Markets", "Smokehouse", "Grill Services"
    ],
    "c_fastfood": [
        "Fast Food", "Chicken Shop", "Chicken Wings", "Donuts", "Buffets", "Poutineries"
    ],
    "c_asian_fusion": [
        "Asian Fusion", "Pan Asian", "Noodles", "Hot Pot", "Dim Sum", "Ramen",
        "Japanese Curry", "Izakaya", "Teppanyaki", "Sushi Bars", "Japanese", "Korean",
        "Chinese", "Thai", "Vietnamese", "Indonesian", "Malaysian", "Singaporean",
        "Filipino", "Burmese", "Cantonese", "Szechuan", "Himalayan/Nepalese", "Taiwanese"
    ],
    "c_bakeries": [
        "Bakeries", "Patisserie/Cake Shop", "Cupcakes", "Macarons", "Bagels",
        "Donuts", "Dessert", "Pretzels", "Creperies"
    ],
    "c_dietary": [
        "Vegetarian", "Vegan", "Gluten-Free", "Halal", "Kosher"
    ]
}


# -----------------------------
# 2. Apply Category Mapping
# -----------------------------
# For each mapped category, create a binary indicator column (1 = category matched)
for col, keywords in category_map.items():
    df_b[col] = df_b["categories"].apply(
        lambda x: int(any(k.lower() in x.lower() for k in keywords))
    )

# Create an "other" category that flags businesses
# which do not belong to any predefined category group
category_cols = list(category_map.keys())
df_b["c_other"] = df_b[category_cols].apply(lambda row: int(row.sum() == 0), axis=1)

# Preview the result
df_b.head()

df_b.to_csv('../datasets/business5.csv', index=False)
